agent-schema/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
requests
python-dotenv
rich
pendulum
google-generativeai
sqlalchemy 
pymysql
cryptography
